DOAC_Script_Generator/requirements.txt
yt-dlp
openai-whisper
torch --index-url https://download.pytorch.org/whl/cu118 